+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/readers/TableKey.java
@@ -3,4 +3,12 @@ package org.openprovenance.prov.service.readers;
 public class TableKey {
     public String isA;
     public Integer ID;
+
+    @Override
+    public String toString() {
+        return "TableKey{" +
+                "isA='" + isA + '\'' +
+                ", ID=" + ID +
+                '}';
+    }
 }