+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/readers/TableKeyList.java
@@ -6,4 +6,10 @@ import java.util.List;
 public class TableKeyList {
     public List<TableKey> key=new LinkedList<>();
 
+    @Override
+    public String toString() {
+        return "TableKeyList{" +
+                "key=" + key +
+                '}';
+    }
 }